add record transaction

+++ src/main/java/org/codewith3h/finmateapplication/repository/AdminLogRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdminLogRepository extends JpaRepository<AdminLog, Long> {
+public interface AdminLogRepository extends JpaRepository<AdminLog, Integer> {
 }
