+++ src/main/java/org/codewith3h/finmateapplication/repository/FinancialReportRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FinancialReportRepository extends JpaRepository<FinancialReport,Long> {
+public interface FinancialReportRepository extends JpaRepository<FinancialReport,Integer> {
 }
