+++ src/main/java/org/codewith3h/finmateapplication/repository/GoalContributionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GoalContributionRepository extends JpaRepository<GoalContribution, Long> {
+public interface GoalContributionRepository extends JpaRepository<GoalContribution, Integer> {
 }
