+++ src/main/java/org/codewith3h/finmateapplication/repository/PremiumPackageRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PremiumPackageRepository extends JpaRepository<PremiumPackage,Long> {
+public interface PremiumPackageRepository extends JpaRepository<PremiumPackage,Integer> {
 }
