+++ src/main/java/org/codewith3h/finmateapplication/repository/SubcriptionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubcriptionRepository extends JpaRepository<Subscription, Long> {
+public interface SubcriptionRepository extends JpaRepository<Subscription, Integer> {
 }
